Move `layers` out of `utils`
