+++ src/otx/algo/detection/backbones/cspnext.py
@@ -13,7 +13,7 @@ from torch import Tensor, nn
 from torch.nn.modules.batchnorm import _BatchNorm
 
 from otx.algo.detection.backbones.csp_darknet import SPPBottleneck
-from otx.algo.detection.utils.layer import CSPLayer
+from otx.algo.detection.layers import CSPLayer
 from otx.algo.modules.base_module import BaseModule
 from otx.algo.modules.conv_module import ConvModule
 from otx.algo.modules.depthwise_separable_conv_module import DepthwiseSeparableConvModule
