+++ src/otx/algo/detection/layers/csp_layer.py
@@ -10,7 +10,7 @@ from typing import TYPE_CHECKING
 import torch
 from torch import Tensor, nn
 
-from otx.algo.detection.utils.layer import ChannelAttention
+from otx.algo.detection.layers import ChannelAttention
 from otx.algo.modules.base_module import BaseModule
 from otx.algo.modules.conv_module import ConvModule
 from otx.algo.modules.depthwise_separable_conv_module import DepthwiseSeparableConvModule
